Move auth domain logic to nauro.auth
The auth command module held both the Typer CLI surface and the auth
domain logic that store/ and sync/ depend on, so library layers imported
from cli/commands/. Move the domain half to a top-level nauro/auth.py:
Auth0 defaults and their env/config overrides, the typed auth errors, the
429 retry helper, config resolution, token load, the single-flight
refresh, the 401 retry wrapper, and the JWT payload decoder.

cli/commands/auth.py keeps only the Typer app, PKCE, the loopback
callback server, and login/status/logout, and imports what it needs from
nauro.auth. No re-export shim is left behind; every importer now points at
nauro.auth. Names called across module boundaries lose their leading
underscore: resolve_auth_config, post_with_429_retry, decode_jwt_payload.

Pure move plus import updates. Bodies, docstrings and comments are
unchanged.
