+++ packages/nauro/CLAUDE.md
@@ -91,6 +91,7 @@ Generated by `nauro sync` into each associated repo root. Contains the L0 payloa
 
 ```
 src/nauro/
+  auth.py                # Auth0 defaults, token load, single-flight refresh, 401 retry
   cli/
     main.py              # Typer app entry point
     commands/            # one module per command
